#300: Moved the php class UniteCompetence to its own file
 ... to avoid code duplication.
